requirements.txt
gradio>=4.44.0
numpy>=1.24.0
opencv-python-headless>=4.9.0.80
pillow>=10.0.0
torch>=2.2.0
torchvision>=0.17.0
torchxrayvision>=1.3.0

streamlit>=1.37.0
